+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Terraform support via WASM plugin
 - Add Zig support via WASM plugin
+- Recognize `language` field and `reference` extra as valid value
 
 - [0.16.0]